agent/requirements.txt
# =============================================================================
# Cortex AI Agent - Dependencies
# =============================================================================

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Technical Analysis
TA-Lib>=0.4.28

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.43.0
optuna>=3.4.0
mlflow>=2.9.0
onnx>=1.15.0
onnxruntime>=1.16.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
redis>=5.0.0
asyncpg>=0.29.0

# Blockchain (Solana)
# Note: Solana SDK is used in TypeScript (eliza/), not Python
# Python is used for offline training and analysis only

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0

# Monitoring
prometheus-client>=0.19.0

# HuggingFace Hub
huggingface-hub>=0.19.0

# Async & Task Queue
celery>=5.3.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0
PyYAML>=6.0.0
aiohttp>=3.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
